libstdc++: Fix some tests that fail with -fexcess-precision=standard
libstdc++-v3/ChangeLog:

	* testsuite/20_util/duration/cons/2.cc: Use values that aren't
	affected by rounding.
	* testsuite/20_util/from_chars/5.cc: Cast arithmetic result to
	double before comparing for equality.
	* testsuite/20_util/from_chars/6.cc: Likewise.
	* testsuite/20_util/variant/86874.cc: Use values that aren't
	affected by rounding.
	* testsuite/25_algorithms/lower_bound/partitioned.cc: Compare to
	original value instead of to floating-point-literal.
	* testsuite/26_numerics/random/discrete_distribution/cons/range.cc:
	Cast arithmetic result to double before comparing for equality.
	* testsuite/26_numerics/random/piecewise_constant_distribution/cons/range.cc:
	Likewise.
	* testsuite/26_numerics/random/piecewise_linear_distribution/cons/range.cc:
	Likewise.
	* testsuite/26_numerics/valarray/transcend.cc (eq): Check that
	the absolute difference is less than 0.01 instead of comparing
	to two decimal places.
	* testsuite/27_io/basic_istream/extractors_arithmetic/char/01.cc:
	Cast arithmetic result to double before comparing for equality.
	* testsuite/27_io/basic_istream/extractors_arithmetic/char/09.cc:
	Likewise.
	* testsuite/27_io/basic_istream/extractors_arithmetic/char/10.cc:
	Likewise.
	* testsuite/27_io/basic_istream/extractors_arithmetic/wchar_t/01.cc:
	Likewise.
	* testsuite/27_io/basic_istream/extractors_arithmetic/wchar_t/09.cc:
	Likewise.
	* testsuite/27_io/basic_istream/extractors_arithmetic/wchar_t/10.cc:
	Likewise.
	* testsuite/ext/random/hoyt_distribution/cons/parms.cc: Likewise.

+++ libstdc++-v3/testsuite/20_util/from_chars/5.cc
@@ -44,7 +44,7 @@ test01()
       r = std::from_chars(s.data(), s.data() + s.length(), d, fmt);
       VERIFY( r.ec == std::errc::invalid_argument );
       VERIFY( r.ptr == s.data() );
-      VERIFY( d == 3.2 );
+      VERIFY( d == (double) 3.2 );
     }
   }
 
@@ -57,7 +57,7 @@ test01()
       r = std::from_chars(s.data(), s.data() + s.length(), d, fmt);
       VERIFY( r.ec == std::errc::invalid_argument );
       VERIFY( r.ptr == s.data() );
-      VERIFY( d == 3.2 );
+      VERIFY( d == (double) 3.2 );
     }
   }
 
@@ -69,7 +69,7 @@ test01()
 			std::chars_format::scientific);
     VERIFY( r.ec == std::errc::invalid_argument );
     VERIFY( r.ptr == s.data() );
-    VERIFY( d == 3.2 );
+    VERIFY( d == (double) 3.2 );
   }
 
   // patterns that are invalid without the final character
@@ -83,7 +83,7 @@ test01()
       r = std::from_chars(s.data(), s.data() + s.length() - 1, d, fmt);
       VERIFY( r.ec == std::errc::invalid_argument );
       VERIFY( r.ptr == s.data() );
-      VERIFY( d == 3.2 );
+      VERIFY( d == (double) 3.2 );
     }
   }
 }
